new refactor in folder structure and also notes page and create quizz page
