+++ client/src/components/dashboard/activityStats.jsx
@@ -1,4 +1,5 @@
 export default function ActivityStats({ stats }) {
+  console.log(stats)
     return (
       <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-4">
         {stats.map((stat, index) => (
