+++ client/src/components/dashboard/studyActivityChart.jsx
