+++ client/src/config/data.js
@@ -1,0 +1,42 @@
+// Sample data for dashbaord
+export const recentNotes = [
+  {
+    id: 1,
+    title: "Data Structures & Algorithms",
+    subject: "Computer Science",
+    date: "2 days ago",
+  },
+  {
+    id: 2,
+    title: "Organic Chemistry Reactions",
+    subject: "Chemistry",
+    date: "1 week ago",
+  },
+  {
+    id: 3,
+    title: "Calculus II: Integration",
+    subject: "Mathematics",
+    date: "2 weeks ago",
+  },
+];
+
+export const upcomingQuizzes = [
+  {
+    id: 1,
+    title: "Database Systems",
+    date: "Tomorrow, 2:00 PM",
+    questions: 20,
+  },
+  { id: 2, title: "Modern Physics", date: "Friday, 10:00 AM", questions: 15 },
+];
+
+export const subjects = [
+    { id: "all", name: "All Subjects" },
+    { id: "cs", name: "Computer Science" },
+    { id: "math", name: "Mathematics" },
+    { id: "physics", name: "Physics" },
+    { id: "chemistry", name: "Chemistry" },
+    { id: "biology", name: "Biology" },
+    { id: "CE", name: "civil engineering" },
+    { id: "ME", name: "mechanical engineering" },
+]
